+++ tests/functional_tests/fcollect_tester.cpp
@@ -141,7 +141,7 @@ template <typename T1>
 void FcollectTester<T1>::resetBuffers(uint64_t size) {
   int n_pes = rocshmem_team_n_pes(ROCSHMEM_TEAM_WORLD);
   for (int i = 0; i < n_pes; i++) {
-    for (int j = 0; j < size; j++) {
+    for (uint64_t j = 0; j < size; j++) {
       // Note: This is redundant work,
       // source is being reinitialized multiple times
       init_buf(source_buf[j], dest_buf[i * size + j]);
@@ -153,10 +153,10 @@ template <typename T1>
 void FcollectTester<T1>::verifyResults(uint64_t size) {
   int n_pes = rocshmem_team_n_pes(ROCSHMEM_TEAM_WORLD);
   for (int i = 0; i < n_pes; i++) {
-    for (int j = 0; j < size; j++) {
+    for (uint64_t j = 0; j < size; j++) {
       auto r = verify_buf(dest_buf[i * size + j], i);
       if (r.first == false) {
-        fprintf(stderr, "Data validation error at idx %d\n", j);
+        fprintf(stderr, "Data validation error at idx %lu\n", j);
         fprintf(stderr, "%s.\n", r.second.c_str());
         // exit(-1);
         return;
