+++ tests/functional_tests/primitive_mr_tester.cpp
@@ -100,9 +100,9 @@ void PrimitiveMRTester::verifyResults(uint64_t size) {
           : 1;
 
   if (args.myid == check_id) {
-    for (int i = 0; i < size; i++) {
+    for (uint64_t i = 0; i < size; i++) {
       if (r_buf[i] != '0') {
-        fprintf(stderr, "Data validation error at idx %d\n", i);
+        fprintf(stderr, "Data validation error at idx %lu\n", i);
         fprintf(stderr, "Got %c, Expected %c\n", r_buf[i], '0');
         exit(-1);
       }
